Break poll chart filters into reusable components
