+++ tracpro/polls/views.py
@@ -40,7 +40,8 @@ class PollCRUDL(smartmin.SmartCRUDL):
 
         def get(self, request, *args, **kwargs):
             self.object = self.get_object()
-            self.filter_form = forms.ChartFilterForm(org=self.object.org, data=request.GET)
+            self.filter_form = forms.PollChartFilterForm(
+                org=self.object.org, data=request.GET)
             return self.render_to_response(self.get_context_data(
                 object=self.object,
                 filter_form=self.filter_form,
